requirements.txt
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
constantly==15.1.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Script==2.0.6
Flask-Twisted==0.1.2
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
mysql-connector-python==8.0.21
observable==1.0.3
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
six==1.15.0
SQLAlchemy==1.3.18
Twisted==20.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.1.0
